Acme.Common/StringHandler.cs
using System;

namespace Acme.Common
{
    public static class StringHandler
    {
        public static string InsertSpaces(string source)
        {
            string result = string.Empty;

            if( !String.IsNullOrEmpty(source))
            {
                foreach (char letter in source)
                {
                    if(char.IsUpper(letter))
                    {
                        result = result.Trim();
                        result += " ";
                    }
                    result += letter;
                }
            }
            result = result.Trim();
            return result;
        }

        public static string InsertSpace(this string source)
        {
            string result = string.Empty;

            if( !String.IsNullOrEmpty(source))
            {
                foreach (char letter in source)
                {
                    if(char.IsUpper(letter))
                    {
                        result = result.Trim();
                        result += " ";
                    }
                    result += letter;
                }
            }
            result = result.Trim();
            return result;
        }
    }
}